backend/requirements.txt
fastapi
uvicorn
pymongo
python-dotenv
pandas
requests
sentence-transformers
faiss-cpu
apscheduler
reportlab
openpyxl
scikit-learn
